// frontend/src/generated/model/teacherActionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherActionResponsePayload } from "./teacherActionResponsePayload";

export interface TeacherActionResponse {
  action_id: string;
  action_type: string;
  applied_at: string;
  payload: TeacherActionResponsePayload;
  reverted_at: string | null;
  status: string;
  version: number;
}