// frontend/src/generated/model/teacherConversationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherMessageResponse } from "./teacherMessageResponse";

export interface TeacherConversationResponse {
  conversation_id: string;
  created_at: string;
  messages: TeacherMessageResponse[];
  profile_id: string;
  status: string;
  title: string;
  updated_at: string;
  version: number;
}